refactor(modules): retire resolver.ts
Nothing in resolver.ts resolved anything after resolveModules was removed;
the name pointed away from the actual solver in resolve/solver.ts.

Move ResolvedModule to resolve/types.ts, next to the ResolveResult that
resolveForApply returns it alongside. Move readModuleConfig to
containers/project/cards-config.ts, next to the readCardsConfig it wraps.
Delete resolver.ts and repoint its nine importers.
