+++ tools/data-handler/src/modules/applier.ts
@@ -20,7 +20,7 @@ import { ProjectPaths } from '../containers/project/project-paths.js';
 import { deleteDir, pathExists } from '../utils/file-utils.js';
 import { getChildLogger } from '../utils/log-utils.js';
 
-import type { ResolvedModule } from './resolver.js';
+import type { ResolvedModule } from './resolve/types.js';
 import type { ModuleSetting } from '../interfaces/project-interfaces.js';
 import type { Project } from '../containers/project.js';
