+++ tools/data-handler/src/modules/resolve/types.ts
@@ -13,7 +13,12 @@
 */
 
 import type { SealFile } from '../../mutations/replay/seal-files.js';
-import type { Source, Version, VersionRange } from '../types.js';
+import type {
+  ModuleDeclaration,
+  Source,
+  Version,
+  VersionRange,
+} from '../types.js';
 
 export type UpdateRequest =
   | { kind: 'verify' }
@@ -47,3 +52,16 @@ export interface ResolveConflict {
 }
 export type ResolveResult =
   { ok: true; changes: Change[] } | { ok: false; conflicts: ResolveConflict[] };
+
+/** A module declaration after the resolver has chosen a concrete version. */
+export interface ResolvedModule {
+  declaration: ModuleDeclaration;
+  /** Git tag/branch to check out; undefined for file sources or unranged git. */
+  ref?: string;
+  /** Remote URL with credentials injected when needed; ready for `SourceLayer.fetch`. */
+  remoteUrl: string;
+  /** Resolved semver version; undefined when no concrete version was pinned. */
+  version?: Version;
+  /** Absolute path where the resolver staged this module's clone. */
+  stagedPath: string;
+}
